Break up objects into files
