+++ Measurement.h
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace Pong
+{
+	class Measurement
+	{
+	public:
+		virtual ~Measurement() {}
+
+		virtual float get_width() const = 0;
+		virtual float get_height() const = 0;
+	};
+}
